Call killpg() with correct pid and only if process hasn't exited.

+++ mininet/node.py
@@ -209,7 +209,7 @@ class Node( object ):
 
     def terminate( self ):
         "Send kill signal to Node and clean up after it."
-        if self.shell:
+        if self.shell and self.shell.returncode is not None:
             os.killpg( self.pid, signal.SIGHUP )
         self.cleanup()
